pyproject.toml
[tool.poetry]
name = "secureml"
version = "0.1.0"
description = "A Python library for privacy-preserving machine learning"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/yourusername/secureml"
documentation = "https://secureml.readthedocs.io"
keywords = ["machine learning", "privacy", "security", "compliance", "gdpr"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
numpy = "^1.24.3"
pandas = "^1.5.3"
syft = "^0.8.1"
opacus = "^1.5.3"
tensorflow-privacy = "^0.8.11"
spacy = "^3.8.3"
faker = "^37.0.1"
pyyaml = "^6.0"
flwr = "^1.12.0"
sdv = "^1.19.0"
jinja2 = "^3.1.4"
matplotlib = "^3.7.1"
click = "^8.1.8"
weasyprint = {version = "^64.1", optional = true}

[tool.poetry.extras]
pdf = ["weasyprint"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
hypothesis = "^6.82.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"
sphinx = "^7.1.0"
sphinx-rtd-theme = "^1.3.0"

[tool.poetry.scripts]
secureml = "secureml.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true 